backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
pillow==10.2.0
numpy==1.26.3
pdf2image==1.17.0
PyPDF2==3.0.1
openpyxl==3.1.2
pydantic==2.5.3
cachetools==5.3.2
